S154 : Header completed + fix nav error if connected
