+++ components/headers/Header.less
@@ -3,9 +3,13 @@
   display: none;
 }
 
-
 // afficher un ombrage lorsque sticky activé
 .uk-sticky.uk-active,
 .ent-header-shadow {
   box-shadow: 0 5px 15px rgba(0, 0, 0, 0.08);
+}
+
+#headerUserDropdown .uk-button-text::before {
+  border: none;
+  transition: none;
 }
